LayoutTests/fast/screen-orientation/orientation-in-resize-event-expected.txt

Checks that screen.orientation.type returns an up-to-date value in the 'resize' event listener.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS wasPortrait is true
Simulating rotation to landscape
* in resize event listener
PASS isLandscape is true
PASS successfullyParsed is true

TEST COMPLETE
